+++ CHANGELOG.rst
@@ -102,6 +102,7 @@ Changed
     - Refactor iRODS ACL constant use (#2225)
     - Remove redundant ``landing_zone_move`` admin access setting task (#2257)
     - Remove ``landing_zone_move`` access cleanup (#2260)
+    - Refactor ``TaskflowProjectTestMixin`` to not require ``login()`` (#2167)
 
 Fixed
 -----